+++ tests/test_macos.py
@@ -1,0 +1,39 @@
+from __future__ import annotations
+
+import os
+from typing import Any
+
+import pytest
+
+from platformdirs.macos import MacOS
+
+
+@pytest.mark.parametrize(
+    "params",
+    [
+        pytest.param({}, id="no_args"),
+        pytest.param({"appname": "foo"}, id="app_name"),
+        pytest.param({"appname": "foo", "version": "v1.0"}, id="app_name_version"),
+    ],
+)
+def test_macos(params: dict[str, Any], func: str) -> None:
+    result = getattr(MacOS(**params), func)
+
+    home = os.path.expanduser("~")
+    suffix_elements = tuple(params[i] for i in ("appname", "version") if i in params)
+    suffix = os.sep.join(("",) + suffix_elements) if suffix_elements else ""
+
+    expected_map = {
+        "user_data_dir": f"{home}/Library/Application Support{suffix}",
+        "site_data_dir": f"/Library/Application Support{suffix}",
+        "user_config_dir": f"{home}/Library/Application Support{suffix}",
+        "site_config_dir": f"/Library/Application Support{suffix}",
+        "user_cache_dir": f"{home}/Library/Caches{suffix}",
+        "user_state_dir": f"{home}/Library/Application Support{suffix}",
+        "user_log_dir": f"{home}/Library/Logs{suffix}",
+        "user_documents_dir": f"{home}/Documents",
+        "user_runtime_dir": f"{home}/Library/Caches/TemporaryItems{suffix}",
+    }
+    expected = expected_map[func]
+
+    assert result == expected